book/requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
cloudinary==1.44.1
Django==5.2.4
django-cloudinary-storage==0.3.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
pillow==11.3.0
psycopg2==2.9.10
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0